requirements.txt
Django==1.8.9
django-filter==0.12.0
djangorestframework==3.3.2
dnspython==1.12.0
Markdown==2.6.5
numpy==1.10.4
Pillow==3.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
requests==2.9.1
sleekxmpp==1.3.1
wheel==0.26.0
